backend/requirements.txt
fastapi[standard]
uvicorn
easyocr
google-genai
python-multipart
pydantic
